requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
python-telegram-bot
